refactor(@eurora/launcher): move current command into a subfolder to prepare for migration
