+++ packages/custom-components/launcher/src/components/bits-command/command-item.svelte
