+++ packages/custom-components/launcher/src/components/bits-command/command-link-item.svelte
