+++ packages/custom-components/launcher/src/components/bits-command/command-shortcut.svelte
